requirements.txt
miniflux
openai
markdownify
markdown
PyYAML
flask
feedgen
schedule
ratelimit
google-genai
Jinja2